Enhance user management: add user ID to UserPublic schema, improve user retrieval by email, and implement JWT token handling for user authentication

+++ backend/api/routers/employee.py
@@ -1,8 +1,9 @@
 from fastapi import APIRouter, HTTPException, Query, Depends
 from typing import Annotated, List
+from backend.api.core.models import User
 from backend.api.services.employee_service import EmployeeService
 from backend.api.core.schemas import EmployeeSchema, EmployeePublic
-from backend.api.dependencies import get_employee_service
+from backend.api.dependencies import get_employee_service, get_current_user
 
 router = APIRouter(
     prefix="/employee",
@@ -12,8 +13,11 @@ router = APIRouter(
 @router.post("", response_model=EmployeePublic)
 def add_employee(
     employee: EmployeeSchema,
-    employee_service: Annotated[EmployeeService, Depends(get_employee_service)]
-):    
+    employee_service: Annotated[EmployeeService, Depends(get_employee_service)],
+    user_logged: User = Depends(get_current_user)
+):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return employee_service.create_employee(employee.name, employee.rg, employee.cpf, employee.role, employee.salary, employee.work_id)
     except Exception as e:
@@ -23,8 +27,11 @@ def add_employee(
 def update_employee(
     id: str,
     employee: EmployeeSchema,
-    employee_service: Annotated[EmployeeService, Depends(get_employee_service)]
+    employee_service: Annotated[EmployeeService, Depends(get_employee_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try: 
         updated_employee = employee_service.update(id, employee.salary, employee.role, employee.work_id)
         if isinstance(updated_employee, str):
@@ -35,8 +42,11 @@ def update_employee(
     
 @router.get("", response_model=List[EmployeePublic])
 def getall_employees(
-    employee_service: Annotated[EmployeeService, Depends(get_employee_service)]
+    employee_service: Annotated[EmployeeService, Depends(get_employee_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return employee_service.all()
     except Exception as e:
@@ -45,8 +55,11 @@ def getall_employees(
 @router.get("/{id}", response_model=EmployeePublic)
 def get_employee(
     id: str,
-    employee_service: Annotated[EmployeeService, Depends(get_employee_service)]
+    employee_service: Annotated[EmployeeService, Depends(get_employee_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         employee = employee_service.get(id)
         if employee is None:
@@ -58,8 +71,11 @@ def get_employee(
 @router.delete("/{id}", response_model=EmployeePublic)
 def delete_employee(
     id:str,
-    employee_service: Annotated[EmployeeService, Depends(get_employee_service)]
+    employee_service: Annotated[EmployeeService, Depends(get_employee_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         result = employee_service.delete(id)
         if result is None:
