+++ backend/api/routers/report.py
@@ -1,9 +1,10 @@
 from fastapi import APIRouter, HTTPException, Depends, Query, Response
 from typing import Annotated, List
+from backend.api.core.models import User
 from backend.api.core.schemas import (PhotoPublic, PhotoSchema, ReportPublic, ReportSchema, ObservationPublic, 
                                   ObservationSchema, ActivityPublic, ActivitySchema, ClimatePublic, ClimateSchema)
 from backend.api.services.report_service import ReportService
-from backend.api.dependencies import get_report_service
+from backend.api.dependencies import get_report_service, get_current_user
 import pandas as pd
 from fastapi.responses import StreamingResponse
 from fpdf import FPDF
@@ -16,8 +17,11 @@ router = APIRouter(
 @router.post("", response_model=ReportPublic)
 def add_report(
     report: ReportSchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return report_service.create_report(report.work_id, report.photos, report.observations, report.activities)
     except Exception as e:
@@ -25,8 +29,11 @@ def add_report(
 
 @router.get("", response_model=List[ReportPublic])
 def getall_reports(
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return report_service.all()
     except Exception as e:
@@ -35,8 +42,11 @@ def getall_reports(
 @router.delete("/{id}", response_model=ReportPublic)
 def delete_report(
     id: str,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return report_service.delete(id)
     except Exception as e:
@@ -45,8 +55,11 @@ def delete_report(
 @router.put("/{id}/addphoto", response_model=PhotoPublic)
 def add_photo(
     photo: PhotoSchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         photo_path = report_service.add_photo(photo.report_id, photo.photo)
         result = PhotoPublic(photo = photo_path)
@@ -57,8 +70,11 @@ def add_photo(
 @router.put("/{id}/removephoto", response_model= PhotoPublic)
 def remove_photo(
     photo: PhotoSchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         photo_path = report_service.remove_photo(photo.report_id, photo.photo)
         return PhotoPublic(photo = photo_path)
@@ -68,8 +84,11 @@ def remove_photo(
 @router.put("/{id}/addobservation", response_model=ObservationPublic)
 def add_observation(
     observation: ObservationSchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         result = report_service.add_observation(observation.report_id, observation.observation)
         return ObservationPublic(observation= result)
@@ -79,8 +98,11 @@ def add_observation(
 @router.put("/{id}/removeobservation", response_model=ObservationPublic)
 def remove_observation(
     observation: ObservationSchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return ObservationPublic(observation= report_service.remove_observation(observation.report_id, observation.observation))
     except Exception as e:
@@ -89,8 +111,11 @@ def remove_observation(
 @router.put("/{id}/addactivity", response_model=ActivityPublic)
 def add_activity(
     activity: ActivitySchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return ActivityPublic(activity= report_service.add_activity(activity.report_id, activity.activity))
     except Exception as e:
@@ -99,8 +124,11 @@ def add_activity(
 @router.put("/{id}/removeactivity", response_model=ActivityPublic)
 def remove_activity(
     activity: ActivitySchema,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return ActivityPublic(activity= report_service.remove_activity(activity.report_id, activity.activity))
     except Exception as e:
@@ -109,8 +137,11 @@ def remove_activity(
 @router.get("/{id}", response_model=ReportPublic)
 def get_report(
     id: str,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return report_service.get(id)
     except Exception as e:
@@ -119,8 +150,11 @@ def get_report(
 @router.get("/{id}/climate")
 def get_climate(
     id: str,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return report_service.get_climate(id)
     except Exception as e:
@@ -128,8 +162,11 @@ def get_climate(
     
 @router.get("/csv/")
 def getall_csv(
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         reports = report_service.all()
         data = []
@@ -149,8 +186,11 @@ def getall_csv(
 @router.get("/{id}/csv")
 def get_csv(
     id: str,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         report = report_service.get(id)
         weather = report_service.get_climate(id)
@@ -172,8 +212,11 @@ def get_csv(
 @router.get("/{id}/pdf")
 def get_pdf(
     id: str,
-    report_service: Annotated[ReportService, Depends(get_report_service)]
+    report_service: Annotated[ReportService, Depends(get_report_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         
         report = report_service.get(id)
