+++ backend/api/routers/work.py
@@ -1,8 +1,9 @@
 from fastapi import APIRouter, Request, Query, Depends, HTTPException
 from typing import Annotated, List
+from backend.api.core.models import User
 from backend.api.services.work_service import WorkService
 from backend.api.core.schemas import WorkSchema, WorkPublic, EmployeePublic, ReportPublic
-from backend.api.dependencies import get_work_service
+from backend.api.dependencies import get_work_service, get_current_user
 
 router = APIRouter(
     prefix="/work",
@@ -12,8 +13,11 @@ router = APIRouter(
 @router.post("", response_model=WorkPublic)
 def add_work(
     work: WorkSchema,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.create_work(work.proprietary_id, work.address)
     except Exception as e:
@@ -21,8 +25,11 @@ def add_work(
     
 @router.get("", response_model=List[WorkPublic])
 def getall_works(
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.all()
     except Exception as e:
@@ -31,8 +38,11 @@ def getall_works(
 @router.get("/{id}", response_model=WorkPublic)
 def get_work(
     id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.get(id)
     except Exception as e:
@@ -41,8 +51,11 @@ def get_work(
 @router.delete("/{id}", response_model=WorkPublic)
 def delete_work(
     id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.delete(id)
     except Exception as e:
@@ -52,8 +65,11 @@ def delete_work(
 def add_report(
     id: str,
     report_id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.add_report(id, report_id)
     except Exception as e:
@@ -63,8 +79,11 @@ def add_report(
 def remove_report(
     id: str,
     report_id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.remove_report(id, report_id)
     except Exception as e:
@@ -73,8 +92,11 @@ def remove_report(
 @router.get("/{id}/proprietary", response_model=WorkPublic)
 def get_proprietary(
     id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.proprietary(id)
     except Exception as e:
@@ -83,8 +105,11 @@ def get_proprietary(
 @router.get("/{id}/reports", response_model=List[ReportPublic])
 def get_reports(
     id: str,
-    work_service: Annotated[WorkService, Depends(get_work_service)]
+    work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.reports(id)
     except Exception as e:
@@ -94,7 +119,10 @@ def get_reports(
 def get_workers(
     id: str,
     work_service: Annotated[WorkService, Depends(get_work_service)],
+    user_logged: User = Depends(get_current_user)
 ):
+    if not user_logged:
+        raise HTTPException(status_code=404, detail="Usuário logado não encontrado.")
     try:
         return work_service.workers(id)
     except Exception as e:
